// src/enums/navigation.ts
export enum Screens {
  // Auth Screens
  INTRO = "Intro",
  GET_STARTED = "GetStarted",
  SIGN_IN = "SignIn",
  SIGN_UP_WITH_EMAIL = "SignUpWithEmail",
  COMPLETE_PROFILE = "CompleteProfile",

  // Onboarding Screens
  ONBOARDING = "Onboarding",
  THANK_YOU = "ThankYou",

  // Settings Screens
  SETTINGS = "Settings",
  PROFILE = "Profile",
  UPDATE_PROFILE = "UpdateProfile",
  CHANGE_PASSWORD = "ChangePassword",
  APPEARANCE = "Appearance",

  // Other Screens
  PRIVACY_POLICY = "PrivacyPolicy",
  TERMS_OF_USE = "TermsOfUse",

  // Modal Screens
  BASE_MODAL = "BaseModal",

  // Chat Screens
  GUEST_CHAT = "GuestChat",
}

export enum Tabs {
  MAIN = "MAIN",
}

export enum MainBottomTabScreens {
  CHAT = "Chat",
  WALLET = "Wallet",
  TRADES = "Trades",
}

export enum Drawers {
  MAIN = "Main",
}